docs(breadcrumb-simple): add dedicated mdx docs for BreadcrumbSimple (#282)
