+++ apps/docs/stories/breadcrumb-simple.mdx
@@ -1,0 +1,141 @@
+import { Meta, Controls, Primary } from '@storybook/addon-docs/blocks';
+import * as BreadcrumbSimpleStories from './breadcrumb-simple.stories';
+
+<Meta of={BreadcrumbSimpleStories} />
+
+# BreadcrumbSimple
+
+Preset that accepts a list of items and renders a navigation breadcrumb with separators, icons, and optional dropdown menus.
+
+For full primitive control, see [Breadcrumb](?path=/docs/primitive-components-breadcrumb-breadcrumb--docs).
+
+## How to use
+
+```tsx
+import { BreadcrumbSimple } from '@signozhq/ui';
+
+export default function MyComponent() {
+	return (
+		<BreadcrumbSimple
+			items={[
+				{ title: 'Home', href: '/' },
+				{ title: 'Products', href: '/products' },
+				{ title: 'Electronics', href: '/products/electronics' },
+				{ title: 'Phones' },
+			]}
+		/>
+	);
+}
+```
+
+<Primary />
+
+## With Icons
+
+Add icons to breadcrumb items using the `icon` prop:
+
+```tsx
+import { BreadcrumbSimple } from '@signozhq/ui';
+import { Home, Folder } from '@signozhq/icons';
+
+<BreadcrumbSimple
+	items={[
+		{ title: 'Home', href: '/', icon: <Home size={16} /> },
+		{ title: 'Documents', href: '/docs', icon: <Folder size={16} /> },
+		{ title: 'Report.pdf' },
+	]}
+/>
+```
+
+## Custom Separator
+
+Change the separator style using the `separator` prop:
+
+```tsx
+// Chevron separator
+<BreadcrumbSimple
+	separator="chevron"
+	items={[
+		{ title: 'Home', href: '/' },
+		{ title: 'Settings', href: '/settings' },
+		{ title: 'Profile' },
+	]}
+/>
+
+// Custom text separator
+<BreadcrumbSimple
+	separator=">"
+	items={[
+		{ title: 'Home', href: '/' },
+		{ title: 'About' },
+	]}
+/>
+```
+
+## With Dropdown Menu
+
+Add dropdown menus to breadcrumb items using the `menu` prop:
+
+```tsx
+<BreadcrumbSimple
+	items={[
+		{ title: 'Home', href: '/' },
+		{
+			title: 'Category',
+			menu: [
+				{ title: 'Electronics', href: '/electronics' },
+				{ title: 'Clothing', href: '/clothing' },
+				{ title: 'Books', href: '/books' },
+			],
+		},
+		{ title: 'Product' },
+	]}
+/>
+```
+
+## With React Router
+
+Use `itemRender` for custom link rendering — useful for integrating with react-router or similar routing libraries:
+
+```tsx
+import { BreadcrumbSimple } from '@signozhq/ui';
+import { Link } from 'react-router-dom';
+
+const items = [
+	{ title: 'Home', path: '/' },
+	{ title: 'Products', path: '/products' },
+	{ title: 'Details', path: '/details' },
+];
+
+<BreadcrumbSimple
+	items={items}
+	itemRender={(route, params, routes, paths) => {
+		const isLast = route === routes[routes.length - 1];
+		return isLast ? (
+			<span>{route.title}</span>
+		) : (
+			<Link to={`/${paths.join('/')}`}>{route.title}</Link>
+		);
+	}}
+/>
+```
+
+## Test IDs
+
+When `testId` is provided, it propagates to all child components:
+
+```tsx
+<BreadcrumbSimple testId="nav" items={items} />
+// Generates:
+// - nav (root)
+// - nav-list
+// - nav-item-0, nav-item-1, ...
+// - nav-link-0, nav-link-1, ...
+// - nav-separator-1, nav-separator-2, ...
+// - nav-page-N (last item)
+// - nav-dropdown-N (if item has menu)
+```
+
+## Props
+
+<Controls of={BreadcrumbSimpleStories.Default} />
